Add shadow and padding to .dropdown-menu__drop
I noticed a dropdown can has the same height as a link on the navigation bar. This creates a dropdown with a low contrast that creates the impression of an overflow issue (z-index).

I added a small padding to the dropdown creating a small offset to the navigation bar. I also added a small shadow to increase the contrast.